added basic support for wikiJSON removal reasons
Still looks for the normal <removalreasons2> tags first.  But if you added the <getfrom> tag it will only look in the wiki.  My reasoning is that since it's a new tag there is no reason to look in the 'old place'.

The other reason is otherwise you could be looking at three $.gets:

1. get <removalreasons> and find a <getfrom> tag.
2. get new subs <removalreasons> and it's empty because they use the wiki.
3. *finally* get new sub's wiki data.

It seems like just making it <getfrom> work using the wiki or not all is the best course.

Another thing that's better about wiki gets is they are much quicker/smaller, because you don't have to download the rest of the stylesheet just to get the reasons.

This can be tested in /r/al_dev, where I direct it back to al_dev using <getfrom>.  Since <getfrom> only supported wiki JSON, it then looks in al_dev's wiki.  Of course you'd never really do that, but it works great for testing.