Bunch of wikiJSON stuff.
Check wiki first for removal reasons.  The reason is that it's *much* quicker to get a 404 than to download the entire stylesheet, find out we don't store the reasons there, *then* check the wiki.

Also switched to escaping/unescaping wiki-stored data to prevent spacing issues with 'compressHTML()'.

As it stands, everything works as expected and pretty quickly in regards to using the wiki and falling back to the stylesheet.  The remaining issues is the ban stuff.  It's so fare and unrelated to the removal code that the only options are to reproduce all of the parsing/fallback code or make the methods *really* generic and place them at the top of the script.